refactor: delete button component, use ui button
