+++ core/app/[locale]/(default)/cart/_components/shipping-options/submit-button.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from 'next-intl';
 import { useFormStatus } from 'react-dom';
 
-import { Button } from '~/components/button';
+import { Button } from '~/components/ui/button';
 
 export const SubmitButton = () => {
   const t = useTranslations('Cart.SubmitShippingCost');
