+++ core/app/[locale]/(default)/compare/_components/add-to-cart.tsx
@@ -3,7 +3,7 @@
 import { useTranslations } from 'next-intl';
 import { useFormStatus } from 'react-dom';
 
-import { Button } from '~/components/button';
+import { Button } from '~/components/ui/button';
 
 export const AddToCart = ({
   disabled = false,
