+++ core/app/[locale]/(default)/login/_components/reset-password-form/index.tsx
@@ -6,7 +6,7 @@ import { useFormStatus } from 'react-dom';
 import ReCaptcha from 'react-google-recaptcha';
 
 import { type FragmentOf } from '~/client/graphql';
-import { Button } from '~/components/button';
+import { Button } from '~/components/ui/button';
 import {
   Field,
   FieldControl,
