+++ core/app/[locale]/(default)/webpages/contact/[id]/contact-us/index.tsx
@@ -6,7 +6,7 @@ import { ChangeEvent, useRef, useState } from 'react';
 import { useFormStatus } from 'react-dom';
 import ReCaptcha from 'react-google-recaptcha';
 
-import { Button } from '~/components/button';
+import { Button } from '~/components/ui/button';
 import {
   Field,
   FieldControl,
